Apply suggestions from code review

+++ sql-files/upgrades/upgrade_20230223.sql
@@ -1,3 +1,3 @@
 ALTER TABLE `db_roulette`
-	ADD COLUMN `chance` int(10) unsigned NOT NULL default '1'
+	ADD COLUMN `chance` tinyint(5) NOT NULL default '1'
 ;
